Working Example with Rock's fixes
- Updates based on pairing with rock
