+++ package.json
@@ -16,28 +16,28 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@flatfile/api": "^1.4.0",
+    "@flatfile/api": "^1.4.8",
     "@flatfile/hooks": "^1.3.0",
     "@flatfile/listener": "^0.3.0",
     "@flatfile/plugin-record-hook": "^0.1.0",
-    "@flatfile/plugin-xlsx-extractor": "^0.0.6",
+    "@flatfile/plugin-xlsx-extractor": "^1.0.0",
     "actions": "^1.3.0",
     "ansi-colors": "^4.1.3",
     "axios": "^1.4.0",
     "chalk": "^5.2.0",
     "email-validator": "^2.0.4",
-    "flatfile": "^3.4.2",
-    "got": "^12.6.0",
-    "moment": "^2.29.4",
-    "node-fetch": "^3.3.1",
+    "flatfile": "^3.4.4",
     "fp-ts": "^2.15.0",
-    "io-ts": "^2.2.20"
+    "got": "^12.6.0",
+    "io-ts": "^2.2.20",
+    "moment": "^2.29.4",
+    "node-fetch": "^3.3.1"
   },
   "devDependencies": {
-    "@types/node": "^18.16.0",
     "@babel/preset-env": "^7.21.5",
     "@babel/preset-typescript": "^7.21.5",
     "@types/jest": "^29.5.1",
+    "@types/node": "^18.16.0",
     "babel-jest": "^29.5.0",
     "esm": "^3.2.25",
     "jest": "^29.5.0",
